Correccions Sprint 2 d'albarans de proveïdor i filtres de cerca #14 #13 #16 #17 #19

See merge request XxericdlxX_2/projecte_4_easytraza_eric!18

+++ backend/src/main/java/cat/copernic/easytraza_backend/dto/mobile/MobileLotSaveRequestDto.java
@@ -3,7 +3,7 @@ package cat.copernic.easytraza_backend.dto.mobile;
 public class MobileLotSaveRequestDto {
 
     private String codiLot;
-    private Integer quantitat;
+    private Double quantitat;
     private String materiaPrimaNom;
 
     public String getCodiLot() {
@@ -14,11 +14,11 @@ public class MobileLotSaveRequestDto {
         this.codiLot = codiLot;
     }
 
-    public Integer getQuantitat() {
+    public Double getQuantitat() {
         return quantitat;
     }
 
-    public void setQuantitat(Integer quantitat) {
+    public void setQuantitat(Double quantitat) {
         this.quantitat = quantitat;
     }
